[backend]: some fixes in BookReview and adjusted some of its endpoints

+++ BookMark.backend/BookMark.src/Controllers/UserController.cs
@@ -14,6 +14,7 @@ using BookMark.Models.Domain;
 using BookMark.Data.Repositories;
 using BookMark.Models.Relationships;
 using BookMark.Models;
+using Microsoft.EntityFrameworkCore;
 
 namespace BookMark.Controllers;
 
@@ -49,34 +50,34 @@ public class UserController : BaseController<User, UserCreateDTO, UserUpdateDTO,
         if (!await _roleManager.RoleExistsAsync(UserRoles.RegularUser))
             await _roleManager.CreateAsync(new IdentityRole(UserRoles.RegularUser));
 
-        var userExists = await _userManager.FindByNameAsync(creationData.Username);
-        if (userExists != null)
-            return Problem(title: "Conflict",
-                            detail: "The username is already taken. Please choose a different one.",
-                            statusCode: StatusCodes.Status409Conflict);
+        var usernameExists = await _userManager.Users.AnyAsync(u => u.UserName == creationData.Username);
+        var emailExists = await _userManager.Users.AnyAsync(u => u.Email == creationData.Email);
+
+        if (usernameExists)
+            return Problem(title: "Conflict", detail: "Try using a different username.", statusCode: StatusCodes.Status409Conflict);
+
+        if (emailExists)
+            return Problem(title: "Conflict", detail: "Try using a different e-mail.", statusCode: StatusCodes.Status409Conflict);
 
         var userToCreate = _mapper.Map<User>(creationData);
         userToCreate.SecurityStamp = Guid.NewGuid().ToString();
 
-        var createdUser = await _userManager.CreateAsync(userToCreate, creationData.Password);
-        if (!createdUser.Succeeded)
+        var creation = await _userManager.CreateAsync(userToCreate, creationData.Password);
+        if (!creation.Succeeded)
             return Problem(title: "User Creation Failed",
                             detail: "An error occurred while creating the user. Please try again later.",
                             statusCode: StatusCodes.Status500InternalServerError);
 
-
         await _userManager.AddToRoleAsync(userToCreate, UserRoles.RegularUser);
 
-        var response = _mapper.Map<UserResponseDTO>(userToCreate);
-
-        return CreatedAtAction(nameof(Get), new { id = userToCreate.Id }, response);
+        var createdUser = await _repository.GetByIdAsync<UserResponseDTO>(userToCreate.Id);
+        return CreatedAtAction(nameof(Get), new { id = userToCreate.Id }, createdUser);
     }
 
 
     [HttpPost("signin")]
     public async Task<IActionResult> SignIn([FromBody] UserCredentialsDTO credentials)
     {
-
         var user = await _userManager.FindByNameAsync(credentials.UsernameOrEmail) ??
                     await _userManager.FindByEmailAsync(credentials.UsernameOrEmail);
 
@@ -88,7 +89,7 @@ public class UserController : BaseController<User, UserCreateDTO, UserUpdateDTO,
         var userRoles = await _userManager.GetRolesAsync(user);
         var authClaims = new List<Claim>
         {
-            new Claim("username", user.UserName!),
+            new Claim(JwtRegisteredClaimNames.Sub, user.Id),
             new Claim(JwtRegisteredClaimNames.Jti, Guid.NewGuid().ToString()),
         };
 
@@ -115,7 +116,7 @@ public class UserController : BaseController<User, UserCreateDTO, UserUpdateDTO,
         var token = new JwtSecurityToken(
                 issuer: _configuration["JWT:ValidIssuer"],
                 audience: _configuration["JWT:ValidAudience"],
-                expires: DateTime.Now.AddHours(4),
+                expires: DateTime.Now.AddMinutes(15),
                 claims: authClaims,
                 signingCredentials: new SigningCredentials(authSigningKey, SecurityAlgorithms.HmacSha256));
 
@@ -123,10 +124,18 @@ public class UserController : BaseController<User, UserCreateDTO, UserUpdateDTO,
     }
 
 
-    [Authorize(Roles = UserRoles.Admin)]
+    [Authorize(Roles = UserRoles.RegularUser)]
     [HttpDelete("delete/{id}")]
-    public override async Task<ActionResult> Delete(string id)
+    public override async Task<ActionResult> Delete([FromRoute] string id)
     {
+        var currentUserId = User.FindFirst(JwtRegisteredClaimNames.Sub)?.Value;
+        var isAdmin = User.IsInRole(UserRoles.Admin);
+
+        if (currentUserId != id && !isAdmin)
+            return Problem(title: "Forbidden",
+                            detail: "You are not authorized to delete this user account.",
+                            statusCode: StatusCodes.Status403Forbidden);
+
         var user = await _userManager.FindByIdAsync(id);
         if (user == null)
             return Problem(title: "Bad Request",
@@ -136,41 +145,58 @@ public class UserController : BaseController<User, UserCreateDTO, UserUpdateDTO,
         var userToDelete = await _userManager.DeleteAsync(user);
         if (!userToDelete.Succeeded)
             return Problem(title: "Unable to Delete User",
-                            detail: "The system encountered an unexpected issue while trying to delete the user account. Please try again or contact support.",
+                            detail: "The system encountered an unexpected issue while trying to delete the user. Please try again or contact support.",
                             statusCode: StatusCodes.Status500InternalServerError);
 
         return NoContent();
     }
 
+
     [Authorize(Roles = UserRoles.RegularUser)]
-    public async Task<ActionResult<BookReviewResponseDTO>> CreateBookReview(BookReviewCreateDTO review)
+    [HttpPost("create-book-review")]
+    public async Task<ActionResult<BookReviewResponseDTO>> CreateBookReview([FromBody] BookReviewCreateDTO review)
     {
+        var currentUserId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+        if (currentUserId == null)
+            return Problem(title: "Unauthorized",
+                            detail: "Invalid session.",
+                            statusCode: StatusCodes.Status401Unauthorized);
+
         if (review.Rating == null && string.IsNullOrWhiteSpace(review.Content))
             return Problem(title: "Bad Request",
-                            detail: "At least one of Rating, Title, or Content must be provided to log a review.",
+                            detail: "Either Rating or Review Content must be provided to log a review.",
                             statusCode: StatusCodes.Status400BadRequest);
 
         var reviewToCreate = _mapper.Map<BookReview>(review);
-
+        reviewToCreate.UserId = currentUserId;
         await _bookReviewRepository.CreateBookReviewAsync(reviewToCreate);
 
-        var createdReview = _bookReviewRepository.GetBookReviewByUser(review.UserId, review.BookId);
-
-        return CreatedAtAction(nameof(GetBookReviewByUser), new { userId = review.UserId, bookId = review.BookId }, createdReview);
+        var createdReview = await _bookReviewRepository.GetBookReviewByCurrentUserAsync(currentUserId, review.BookId);
+        return CreatedAtAction(nameof(GetBookReviewByCurrentUser), new { userId = currentUserId, bookId = review.BookId }, createdReview);
     }
 
+
     [Authorize(Roles = UserRoles.RegularUser)]
-    public async Task<ActionResult<BookReviewResponseDTO?>> GetBookReviewByUser(string userId, string bookId)
+    [HttpGet("get-book-review-by-current-user/{bookId}")]
+    public async Task<ActionResult<BookReviewResponseDTO?>> GetBookReviewByCurrentUser([FromRoute] string bookId)
     {
-        var review = await _bookReviewRepository.GetBookReviewByUser(userId, bookId);
+        var currentUserId = User.FindFirst(JwtRegisteredClaimNames.Sub)?.Value;
+        if(currentUserId == null)
+            return Problem(title: "Unauthorized",
+                            detail: "Invalid session.",
+                            statusCode: StatusCodes.Status401Unauthorized);
+
+        var review = await _bookReviewRepository.GetBookReviewByCurrentUserAsync(currentUserId, bookId);
 
         return Ok(review);
     }
 
-    public async Task<ActionResult<BookReviewStatsDTO>> GetBookReviewStats(string bookId)
+
+    [HttpGet("get-book-review-stats/{bookId}")]
+    public async Task<ActionResult<BookReviewStatsDTO>> GetBookReviewStats([FromRoute] string bookId)
     {
-        var averageRating = await _bookReviewRepository.GetBookReviewAverageRating(bookId);
-        var reviewCount = await _bookReviewRepository.GetBookReviewCount(bookId);
+        var averageRating = await _bookReviewRepository.GetBookReviewAverageRatingAsync(bookId);
+        var reviewCount = await _bookReviewRepository.GetBookReviewCountAsync(bookId);
 
         var stats = new BookReviewStatsDTO
         {
@@ -181,15 +207,36 @@ public class UserController : BaseController<User, UserCreateDTO, UserUpdateDTO,
         return Ok(stats);
     }
 
-    public async Task<ActionResult<Page<BookReviewResponseDTO>>> GetLatestBookReviews(string bookId, int pageIndex, int pageSize)
+
+    [HttpGet("get-latest-book-reviews/{bookId}")]
+    public async Task<ActionResult<Page<BookReviewResponseDTO>>> GetLatestBookReviews([FromRoute] string bookId, [FromQuery] int pageIndex, [FromQuery] int pageSize)
     {
-        var page = await _bookReviewRepository.GetLatestBookReviews(bookId, pageIndex, pageSize);
+        var page = await _bookReviewRepository.GetLatestBookReviewsAsync(bookId, pageIndex, pageSize);
 
         return Ok(page);
     }
 
-    public async Task<ActionResult> DeleteBookReview(string userId, string bookId)
+    [HttpGet("get-latest-book-reviews-by-user/{userId}")]
+    public async Task<ActionResult<Page<BookReviewResponseDTO>>> GetLatestBookReviewsByUser([FromRoute] string userId, [FromQuery] int pageIndex, [FromQuery] int pageSize)
     {
+        var page = await _bookReviewRepository.GetLatestBookReviewsByUserAsync(userId, pageIndex, pageSize);
+
+        return Ok(page);
+    }
+
+
+    [Authorize(Roles = UserRoles.RegularUser)]
+    [HttpDelete("delete-book-review/{userId}/{bookId}")]
+    public async Task<ActionResult> DeleteBookReview([FromRoute] string userId, [FromRoute] string bookId)
+    {
+        var currentUserId = User.FindFirst(JwtRegisteredClaimNames.Sub)?.Value;
+        var isAdmin = User.IsInRole(UserRoles.Admin);
+
+        if (currentUserId != userId && !isAdmin)
+            return Problem(title: "Forbidden",
+                            detail: "You are not authorized to delete this book review.",
+                            statusCode: StatusCodes.Status403Forbidden);
+
         await _bookReviewRepository.DeleteBookReviewAsync(userId, bookId);
 
         return NoContent();
